doc: Fix TOC generation for refentries
The XSLT stylesheets by default don't show refentries inside sections in
table of contents, which for our documentation leads to some regressions
compared to the DSSSL output.  For example, in the SPI chapter, which is
mostly refentries, you don't get any usable table of contents.  Tweak
things so it's mostly back to what it was before.